\name{example_data}
\alias{example_data}
\docType{data}
\title{
Example data, 50 samples, 22600 genes.
}
\description{
Example data, consisting of 50 samples, 22600 genes. Example data were randomly generated. Disease data such as this example data should be formatted either as a data frame or ExpressionSet. Data frames should have rows corresponding to patients and columns as genes with Entrez ID column names. ExpressionSet probes should be mapped to Entrez ID first before proceeding.
}
\usage{data("example_data")}
