\name{example_healthy}
\alias{example_healthy}
\docType{data}
\title{
Example healthy data, 100 samples, 22600 genes.
}
\description{
Example healthy data, 100 samples, 22600 genes. Example healthy data were randomly generated. Healthy data such as this example healthy data should be formatted either as a data frame or ExpressionSet. Data frames should have rows corresponding to patients and columns as genes with Entrez ID column names. ExpressionSet probes should be mapped to Entrez ID first before proceeding.
}
\usage{data("example_healthy")}
